Fix return in BCC-Examples

+++ BCC-Examples/hello_fields.py
@@ -6,7 +6,7 @@ from ctypes import c_void_p, c_int64
 @section("tracepoint/syscalls/sys_enter_clone")
 def hello_world(ctx: c_void_p) -> c_int64:
     print("Hello, World!")
-    return c_int64(0)
+    return 0  # type: ignore [return-value]
 
 
 @bpf
